lab fucntionality complete
